+++ CHANGELOG.rst
@@ -4,8 +4,10 @@ CHANGELOG
 
 
 1.18.2.dev
-======
+==========
+
 * enhancement: Include SageMaker Notebook Instance version number in boto3 user agent, if available.
+* feature: Support for updating existing endpoint
 
 1.18.1
 ======